requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
flask>=2.3.0
python-dotenv>=1.0.0
pytz>=2023.3
APScheduler>=3.10.0
